// src/services/client/Clientes.ts
export type Clientes = {
  _id: string;
  nomecompleto?: string;
  apelido?: string;
  bairro?: string;
  cep?: string;
  cgc: string;
  cidade?: string;
  contrato?: boolean;
  fiscal?: string;
  pessoal?: string;
  contabil?: string;
  email?: string;
  estado?: string;
  pais?: string;
  fantasia?: string;
  ramodeatividade?: string;
  logradouro?: string;
  numero?: string;
  passsefaz?: string;
  plcontabil?: boolean;
  plfiscal?: boolean;
  plpessoal?: boolean;
  ramoatividade?: string;
  razaosocial?: string;
  telefone?: string;
  tipopessoa?: string;
  usursefaz?: string;
  cnae?: string;
  ie?: string;
  complemento?: string;
};